pyproject.toml
[tool.poetry]
name = "fastapi-sentry"
version = "0.1.0"
description = "FastAPI with Sentry Practice"
authors = ["weekwith.me <leedobby@weekwith.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.26"
sentry-sdk = "^1.4.3"
asyncpg = "^0.24.0"
alembic = "^1.7.4"
python-dotenv = "^0.19.2"
python-jose = "^3.3.0"
email-validator = "^1.1.3"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
httpx = "^0.20.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
